Created the foundations for the client side chat
ChatSession.js stores all of the information about the current
user and their room.

Message.js is a data structure for storing and converting messages
to html.

ServiceBuilder.js is a platform for preparing JSON objects for
command/service requests.

join.js is the a test run for the connection process.