Re-designed client side chat to make multiple rooms work with one socket
Changed ChatSession.js to Room.js and created a new ChatSession.js to hold an array of rooms, as well as the user and their associated socket object.

Started on the initial launch file (main.js)

Also added a util file with some cheeky monkey-patching for arrays.
(Planning to get this out of here long before release)